pyproject.toml
[project]
name = "sightcall-qa-api"
version = "0.1.0"
description = "A RAG-based API to ask questions about SightCall."
readme = "README.md"
requires-python = ">=3.13,<3.14"
license = "AGPL-3.0-or-later"
authors = [
    {name = "Charles-Meldhine Madi Mnemoi", email = "charlesmeldhine.madimnemoi@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "haystack-ai>=2.1.3",
    "pgvector-haystack>=3.3.0",
]

[project.scripts]
indexation = "sightcall_qa_api.indexation.presentation.cli.indexation:app"

[project.urls]
"Bug Tracker" = "https://github.com/cmnemoi/sightcall_qa_api/issues"
Changelog = "https://github.com/cmnemoi/sightcall_qa_api/blob/main/CHANGELOG.md"
Homepage = "https://github.com/cmnemoi/sightcall_qa_api/"
Repository = "https://github.com/cmnemoi/sightcall_qa_api/"

[dependency-groups]
dev = [
    "pytest-watcher>=0.4.3",
    "python-semantic-release>=9.21.1",
]
indexation = [
    "nltk>=3.9.1",
    "trafilatura>=2.0.0",
    "typer>=0.15.4",
]
lint = [
    "mypy>=1.15.0",
    "pytest-mypy>=1.0.1",
    "ruff>=0.11.10",
]
test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "testcontainers[postgres]>=4.10.0",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sightcall_qa_api"]
